+++ libft/libft.h
@@ -38,7 +38,7 @@ int		ft_isalpha(int c);
 int		ft_isalnum(int c);
 void	*ft_calloc(size_t nitems, size_t size);
 void	ft_bzero(void *s, size_t n);
-int		ft_atoi(const char *str);
+long int	ft_atoi(const char *str);
 char	*ft_strdup(const char *str);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strjoin(char const *s1, char const *s2);
